requirements.txt
# Python Chatbot Requirements
# ----------------------------
# This file lists all the Python packages needed to run the chatbot.
# Install these packages using: pip install -r requirements.txt

# Streamlit - Web framework for creating the UI
streamlit>=1.32.0

# LangChain - Framework for building LLM applications (compatible versions)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Google Generative AI - Official Google library for Gemini API
google-generativeai>=0.8.0

# Python-dotenv - Load environment variables from .env file
python-dotenv>=1.0.0

# Core dependencies
pydantic>=2.5.0

# Database - PostgreSQL
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Authentication
bcrypt>=4.0.0
